Initial commit.
shower.py and shower_test.py implement the bellman eqn for
the shower temperature control environment from scratch.

shower_rl_example notebook tries to do the same with open-ai.
Currently, it runs but the model fails badly.

Also a yml file for the open ai conda environment. And just
some notes in install_history to recreate the conda env
because keral-rl2 is needed, but must be installed from src.